Support saving crash dumps on systems hosted in RAM
By default, Kdump copies the entire vmcore to the disk and reboots the
system afterwards. While this works in most cases, it does not work for
systems hosted in RAM where the crash dump will be lost once rebooted.
To support this scenario, a new option in the kdump.conf,
save_core_to_ram, is added and commented out. Setting this option to 1
will allow a compressed crash dump to be produced by utilizing the
core_collector option if specified, otherwise the dmesg log will be
extracted from the vmcore. This is especially necessary because there is
no disk to save the dump to, and there is not enough space in the
pre-allocated crashkernel memory space to save an entire vmcore.